84-87

+++ lib/include/prime.h
@@ -7,7 +7,7 @@ private:
 public:
     PrimeIterator();
     PrimeIterator(uint64_t max);
-    unsigned int operator*() const;
+    uint64_t operator*() const;
     PrimeIterator& operator++();
     bool operator!=(const PrimeIterator& other) const;
     PrimeIterator operator+(int offset);
